feat(dashboard/tenants): surface proxy public URL banner (closes #251)
Tenant operators landing on /tenants previously had no way to discover
the proxy's public URL — the address their applications must POST
/v1/* traffic to. The only source of that URL was the original Basilica
provision result JSON.

This change adds a "Connection" banner at the top of /tenants showing:
- The proxy URL (read from LLMTRACE_PROXY_URL, which is already
  injected at runtime by deployments/basilica/lifecycle.py via
  _inject_proxy_url_into_dashboard).
- A copy-to-clipboard button for the URL.
- A copy-pasteable curl example with the URL substituted in.
- A yellow warning card when LLMTRACE_PROXY_URL is unset, so operators
  immediately see that the dashboard was provisioned without it.

Because the existing /tenants page is "use client", env vars are not
available there. The page is split:
- dashboard/src/app/tenants/page.tsx becomes a thin server component
  that renders the banner + the existing tenant table client. Marked
  force-dynamic so process.env.LLMTRACE_PROXY_URL is resolved at
  request time, not baked in at build time.
- dashboard/src/app/tenants/_client.tsx holds the unchanged tenant
  table client logic (renamed from TenantsPage to TenantsClient).
- dashboard/src/components/proxy-connection-banner.tsx is the server
  component that reads the env var.
- dashboard/src/components/proxy-connection-banner-client.tsx renders
  the populated banner (interactive copy buttons) and the missing-URL
  notice. Reuses the same clipboard-write fallback pattern the
  existing tenants page uses.

A Playwright smoke test in dashboard/e2e/dashboard.spec.ts asserts
either the populated banner (URL value + copy buttons + curl example)
or the missing-state card renders — covering both deployment shapes.
